Fix test non-determinism: (#52)
The probe command truncates the responses to the probe requests and only keeps the first 10 responses. This can cause test failures as the responses are out of order and particularly in CI the probe-0 would be missing in the first 10 requests.
The fix is to not truncate the responses which can be verbose but it prevents failures such as this one from creeping.

Signed-off-by: Rashid Kaleem <230885705+arekay-nv@users.noreply.github.com>